+++ src/lib/Components/Inbox/Conversations/Preview/ArtworkPreview.tsx
@@ -5,6 +5,8 @@ import { TouchableHighlight } from "react-native"
 
 import { PreviewText as P, Subtitle } from "../../Typography"
 
+import { Schema, Track, track as _track } from "../../../../utils/track"
+
 import OpaqueImageView from "lib/Components/OpaqueImageView"
 import colors from "lib/data/colors"
 import fonts from "lib/data/fonts"
@@ -57,12 +59,26 @@ interface Props extends RelayProps {
   onSelected?: () => void
 }
 
+const track: Track<Props> = _track
+
+@track()
 export class ArtworkPreview extends React.Component<Props, any> {
+  @track((props, state) => ({
+    action_type: Schema.ActionTypes.Tap,
+    action_name: Schema.ActionNames.ConversationAttachmentArtwork,
+    owner_type: Schema.OwnerEntityTypes.Artwork,
+    owner_id: props.artwork._id,
+    owner_slug: props.artwork.id,
+  }))
+  attachmentSelected() {
+    this.props.onSelected()
+  }
+
   render() {
     const artwork = this.props.artwork
 
     return (
-      <TouchableHighlight underlayColor={colors["gray-light"]} onPress={this.props.onSelected}>
+      <TouchableHighlight underlayColor={colors["gray-light"]} onPress={() => this.attachmentSelected()}>
         <Container>
           <Image imageURL={artwork.image.url} />
           <TextContainer>
@@ -86,6 +102,8 @@ export default createFragmentContainer(
   ArtworkPreview,
   graphql`
     fragment ArtworkPreview_artwork on Artwork {
+      id
+      _id
       title
       artist_names
       date
@@ -98,6 +116,8 @@ export default createFragmentContainer(
 
 interface RelayProps {
   artwork: {
+    id: string | null
+    _id: string | null
     title: string | null
     artist_names: string | null
     date: string | null